src/test/java/com/sbrf/reboot/calculator/CalculatorTest.java
package com.sbrf.reboot.calculator;

import org.junit.jupiter.api.Test;

import static org.junit.jupiter.api.Assertions.*;

class CalculatorTest {

    @Test
    void getAddition() {
        assertEquals(9, Calculator.getAddition(4, 5));
    }

    @Test
    void getSubtraction() {
        assertEquals(-1, Calculator.getSubtraction(4, 5));
    }

    @Test
    void getMultiplication() {
        assertEquals(20, Calculator.getMultiplication(4, 5));
    }

    @Test
    void getDivision() {
        assertEquals(3, Calculator.getDivision(9, 3));
    }

    @Test
    void getPow() {
        assertEquals(1024, Calculator.getPow(2, 10));
    }

    @Test
    void getFactorial() {
        assertEquals(120, Calculator.getFactorial(5));
    }

    @Test
    void getPercent() {
        assertEquals(0.05, Calculator.getPercent(5));
    }

    @Test
    void classHasSevenMethods(){
        assertEquals(7,Calculator.class.getMethods().length-Object.class.getMethods().length);
    }

}